avformat/hls: avoid seeking to the very end of segments
If seeking is requested to the very end of a segment (e.g. less than a
second from the end), the segment still has to be downloaded in full, and
almost immediately afterwards the next segment will have to be
downloaded.

To avoid that, skip segments when seeking to a point less than 100ms
from the end of a segment.

Signed-off-by: Anssi Hannula <anssi.hannula@iki.fi>